+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/openai-compatible
 
+## 3.0.0-canary.45
+
+### Patch Changes
+
+- Updated dependencies [ca446f8]
+  - @ai-sdk/provider-utils@5.0.0-canary.38
+
+## 3.0.0-canary.44
+
+### Patch Changes
+
+- Updated dependencies [d848405]
+  - @ai-sdk/provider-utils@5.0.0-canary.37
+
+## 3.0.0-canary.43
+
+### Patch Changes
+
+- Updated dependencies [ca39020]
+  - @ai-sdk/provider-utils@5.0.0-canary.36
+
 ## 3.0.0-canary.42
 
 ### Patch Changes