+++ packages/otel/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @ai-sdk/otel
 
+## 1.0.0-canary.77
+
+### Patch Changes
+
+- Updated dependencies [b67525f]
+- Updated dependencies [ca446f8]
+- Updated dependencies [bcacd48]
+  - ai@7.0.0-canary.131
+
+## 1.0.0-canary.76
+
+### Patch Changes
+
+- ai@7.0.0-canary.130
+
+## 1.0.0-canary.75
+
+### Patch Changes
+
+- Updated dependencies [d1b3786]
+  - ai@7.0.0-canary.129
+
+## 1.0.0-canary.74
+
+### Patch Changes
+
+- ai@7.0.0-canary.128
+
+## 1.0.0-canary.73
+
+### Patch Changes
+
+- c025d60: feat(otel): add option for custom span attributes
+- Updated dependencies [e95e38d]
+- Updated dependencies [016e877]
+- Updated dependencies [ca99fea]
+- Updated dependencies [d775a57]
+- Updated dependencies [538c12b]
+  - ai@7.0.0-canary.127
+
 ## 1.0.0-canary.72
 
 ### Patch Changes