+++ packages/voyage/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/voyage
 
+## 2.0.0-canary.13
+
+### Patch Changes
+
+- Updated dependencies [ca446f8]
+  - @ai-sdk/provider-utils@5.0.0-canary.38
+
+## 2.0.0-canary.12
+
+### Patch Changes
+
+- Updated dependencies [d848405]
+  - @ai-sdk/provider-utils@5.0.0-canary.37
+
+## 2.0.0-canary.11
+
+### Patch Changes
+
+- Updated dependencies [ca39020]
+  - @ai-sdk/provider-utils@5.0.0-canary.36
+
 ## 2.0.0-canary.10
 
 ### Patch Changes